Breaking Changes
chore(deps): upgraded dependencies
fix(bugs): fixed bugs in backend
refactor: client codebase refactored
