--- client/src/input.css
@@ -1,13 +1,0 @@
-@tailwind base;
-@tailwind components;
-
-@layer components{
-    .sidebar{
-        @apply lg:fixed absolute z-50 left-0 top-0 h-screen bg-white dark:bg-gray-900 dark:text-white lg:p-1 p-2  transition-[left] duration-1000 ease-in-out;
-    }
-    .side-nav-btn{
-        @apply flex items-center justify-center font-bold xl:px-7 lg:px-6 lg:py-6 h-full text-center w-full bg-white text-gray-900 rounded-lg hover:bg-blue-200 hover:text-blue-700 text-base lg:text-lg p-4;
-    }
-}
-
-@tailwind utilities;
