+++ client/tailwind.config.js
@@ -1,12 +1,9 @@
-module.exports = {
+/** @type {import('tailwindcss').Config} */
+export default {
   content: ["./src/**/*.{html,js,jsx}"],
   theme: {
-    extend: {
-      transitionProperty: {
-        'height': 'height',
-        'spacing': 'margin, padding',
-      }
-    },
+    extend: {},
   },
   plugins: [],
 }
+
